invoice_cataloger/requirements.txt
# PDF Processing (Multiple engines for robustness)
pypdf>=3.17.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0
pikepdf>=8.0.0

# OCR (Multiple engines)
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=1.16.3

# Document Processing
python-docx>=1.1.0
openpyxl>=3.1.0
extract-msg>=0.45.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# AI/LLM Integration
requests>=2.31.0
openai>=1.0.0

# Data Processing & Export
pandas>=2.1.0
xlsxwriter>=3.1.0

# Utilities
python-magic-bin>=0.4.14; platform_system == "Windows"
python-magic>=0.4.27; platform_system != "Windows"
tqdm>=4.66.0
colorama>=0.4.6
python-dateutil>=2.8.2
